+++ src/index.tsx
@@ -1,4 +1,4 @@
-﻿import { Form, TooltipItem, Checkbox, CheckboxList, CustomDatePicker, DatePickerRange, EmailInput, FileUpload, NumberInput, PersonalIdentityInput, Select, Textarea, TextInput, LinkArray, HiddenInput, CustomSubmitButton, CustomSubmitButtonType, RadioButton, FormLink, ClearFormButton } from "./components/Form/Form";
+﻿import { Form, TooltipItem, MultiSelect, Checkbox, CheckboxList, CustomDatePicker, DatePickerRange, EmailInput, FileUpload, NumberInput, PersonalIdentityInput, Select, Textarea, TextInput, LinkArray, HiddenInput, CustomSubmitButton, CustomSubmitButtonType, RadioButton, FormLink, ClearFormButton } from "./components/Form/Form";
 import { Modal, ModalBody, ModalFooter, ModalSize } from "./components/Modal/Modal";
 import { Table } from "./components/Table/Table";
 import { MainNavigation } from "./components/MainNavigation/MainNavigation";
@@ -63,5 +63,6 @@ export {
     SortTableProvider,
     ClearFormButton,
     FormDownloadButton,
-    TooltipItem
+    TooltipItem,
+    MultiSelect
 }